refactor: ui and some codes
