+++ lib/presentation/pages/search/partial/club_card_widget.dart
@@ -1,0 +1,51 @@
+import 'package:flutter/material.dart';
+import 'package:liga_inggris_mobile/presentation/common/club_logo_widget.dart';
+import 'package:liga_inggris_mobile/services/club/model.dart';
+
+class ClubCardWidget extends StatelessWidget {
+  final ClubModel club;
+
+  const ClubCardWidget({super.key, required this.club});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: const EdgeInsets.all(8.0),
+      child: Card(
+        color: Colors.transparent,
+        shape: RoundedRectangleBorder(
+          side: const BorderSide(color: Colors.white, width: 2.0),
+          borderRadius: BorderRadius.circular(8.0),
+        ),
+        child: SizedBox(
+          height: 120,
+          child: Row(
+            crossAxisAlignment: CrossAxisAlignment.center,
+            children: [
+              Padding(
+                padding: const EdgeInsets.only(left: 12),
+                child: Container(
+                  width: 100,
+                  alignment: Alignment.center,
+                  child: ClubLogoWidget(
+                    imageUrl: club.logo ?? '',
+                    width: 80,
+                    height: 80,
+                  ),
+                ),
+              ),
+              Container(
+                alignment: Alignment.centerLeft,
+                padding: const EdgeInsets.only(left: 16.0, right: 32.0),
+                child: Text(
+                  club.name ?? 'Unknown Club',
+                  style: const TextStyle(color: Colors.white),
+                ),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
